+++ bdtsim/protocol/fairswap/fairswap.py
@@ -27,7 +27,8 @@ from bdtsim.account import Account
 from bdtsim.contract import SolidityContract
 from bdtsim.data_provider import DataProvider
 from bdtsim.environment import Environment
-from bdtsim.protocol import Protocol, ProtocolManager, ProtocolInitializationError, ProtocolExecutionError
+from bdtsim.protocol import Protocol, ProtocolManager, ProtocolInitializationError, ProtocolExecutionError,\
+    DEFAULT_ASSET_PRICE
 from bdtsim.protocol_path import ProtocolPath
 from bdtsim.util.bytes import generate_bytes
 from . import merkle, encoding
@@ -106,19 +107,19 @@ class FairSwap(Protocol):
         return SolidityContract(FairSwap.CONTRACT_NAME, contract_code=contract_code_rendered)
 
     def execute(self, protocol_path: ProtocolPath, environment: Environment, data_provider: DataProvider,
-                seller: Account, buyer: Account, price: int = 1000000000) -> None:
+                seller: Account, buyer: Account, price: int = DEFAULT_ASSET_PRICE) -> None:
         """Execute a file transfer using the FairSwap protocol / FileSale contract.
 
         Args:
-            protocol_path (ProtocolPath):
-            environment (Environment):
-            data_provider (DataProvider):
-            seller (Account):
-            buyer (Account):
-            price (int):
+            protocol_path (ProtocolPath): Protocol path this simulation for a data exchange will take
+            environment (Environment): The environment (blockchain) in which the protocol interactions will take place
+            data_provider (DataProvider): The data to be traded
+            seller (Account): The selling party
+            buyer (Account): The buying party
+            price (int): Price for the data/asset to be traded (in Wei)
 
         Returns:
-            None:
+            None
         """
         # initial assignments and checks
         slice_length = int(data_provider.data_size / self.slices_count)